refactor(mitm-addon): split firewall request matching tests
